+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Changed: path(name) of the base url now have a trailing slash
 - Added: `MetadataProvider` to send the data you want to the React context
 in a `metadata` attribute.
+- Added: `__BASE_URL__` is now send as constantes.
+- Added: `statinamic/lib/prepare-defined-values` helper to easily use JS
+object with `webpack.DefinePlugin`
 
 # 0.1.0 - 2015-10-21